+++ blossom_ai/__init__.py
@@ -83,7 +83,7 @@ from blossom_ai.utils import (
     cached,
 )
 
-__version__ = "0.4.2"
+__version__ = "0.4.3"
 
 __all__ = [
     # Main client